+++ dal-cpp/dal/script/visitor/pastevaluator.hpp
@@ -13,8 +13,7 @@ namespace Dal::Script {
     class PastEvaluator_: public EvaluatorBase_<T_, PastEvaluator_> {
     public:
         using Base = EvaluatorBase_<T_, PastEvaluator_>;
-        explicit PastEvaluator_(const Vector_<T_>& variables, const Vector_<T_>& constVariables = Vector_<T_>())
-            : Base(variables, constVariables) {}
+        explicit PastEvaluator_(const Vector_<>& variables, const Vector_<T_>& constVariables = Vector_<T_>()) : Base(variables, constVariables) {}
 
         using Base::Visit;
         using Base::VisitNode;
@@ -32,9 +31,7 @@ namespace Dal::Script {
             Base::Visit(node);
         }
 
-        [[nodiscard]] FORCE_INLINE const Vector_<>& Variables() const {
-            return variables_;
-        }
+        [[nodiscard]] FORCE_INLINE const Vector_<T_>& Variables() const { return variables_; }
     };
 
 } // namespace Dal::Script